requirements.txt
flask
rembg
pillow
waitress
google-generativeai
